add GObject destructor code
Moving the munmap() call into the finalize callback makes sure
that it can't be called until all attached ClaylandSurface objects
have released that particular buffer. Since the CoglHandle depends
on access to that memory, this is a very good thing.

While we're at it, move the cogl_handle_unref call where it
belongs, in the ClaylandBuffer finalize method.