add output object
This is a prerequisite for fullscreen support. Without
an output object, we never sent an output geometry event,
so the clients couldn't know what size a surface would be
when fullscreen.

TODO: Allow multiple outputs per compositor, and tell clients
when an output's geometry changes dynamically, instead of
assuming it's constant.